changelog: Bug Fixes, Remove Unused Providers, fix error encountered when removing providers without integrations

+++ lib/cleanup/destroyable_records.rb
@@ -26,7 +26,7 @@ class DestroyableRecords
 
     stdout.puts '********'
     stdout.puts 'Integration:'
-    if integration.blank?
+    if integration.nil?
       stdout.puts 'No associated integration'
     else
       stdout.puts integration.attributes.to_yaml
@@ -58,15 +58,17 @@ class DestroyableRecords
   end
 
   def destroy_records
-    stdout.puts 'Destroying integration usages'
-    integration_usages.each do |integration_usage|
-      integration_usage.destroy!
-    end
-    integration.reload
+    if integration
+      stdout.puts 'Destroying integration usages'
+      integration_usages.each do |integration_usage|
+        integration_usage.destroy!
+      end
+      integration.reload
 
-    stdout.puts "Destroying integration with issuer #{integration.issuer}"
-    integration.destroy!
-    service_provider.reload
+      stdout.puts "Destroying integration with issuer #{integration.issuer}"
+      integration.destroy!
+      service_provider.reload
+    end
 
     stdout.puts "Destroying service provider issuer #{service_provider.issuer}"
     service_provider.destroy!
@@ -79,11 +81,11 @@ class DestroyableRecords
   private
 
   def integration_usages
-    integration&.integration_usages || []
+    integration&.integration_usages
   end
 
   def iaa_orders
-    integration&.iaa_orders || []
+    integration&.iaa_orders
   end
 
   def in_person_enrollments
